♻️ refactor : 이메일 기능 주석 추가
