+++ src/main/java/com/gdg/homepage/landing/email/domain/repository/VerificationCodeRepository.java
@@ -1,12 +1,12 @@
-package com.gdg.homepage.landing.member.domain.repository;
+package com.gdg.homepage.landing.email.domain.repository;
 
-import com.gdg.homepage.landing.member.domain.entity.VerificationCode;
+import com.gdg.homepage.landing.email.domain.entity.VerificationCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-public interface EmailRepository extends JpaRepository<VerificationCode, Long> {
+public interface VerificationCodeRepository extends JpaRepository<VerificationCode, Long> {
 
     Optional<VerificationCode> findByEmailAndCode(String email, String code);
     void deleteByExpiresTimeBefore(LocalDateTime time);
