+++ src/main/java/com/gdg/homepage/landing/email/presentation/EmailApi.java
@@ -5,9 +5,8 @@ import com.gdg.homepage.core.response.ApiResponse;
 import com.gdg.homepage.landing.email.application.dto.request.EmailSendRequest;
 import com.gdg.homepage.landing.email.application.dto.request.EmailVerifyRequest;
 import com.gdg.homepage.landing.email.application.usecase.EmailUseCase;
-import com.gdg.homepage.landing.email.exception.NotVerifiedException;
+import com.gdg.homepage.landing.email.application.exception.NotVerifiedException;
 import com.gdg.homepage.landing.email.presentation.swagger.EmailApiSpec;
-import io.swagger.v3.oas.annotations.Operation;
 import jakarta.mail.MessagingException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
